Split apart db.js
