+++ package.json
@@ -32,6 +32,7 @@
     "mime": "^2.4.4",
     "multer": "^1.4.2",
     "multer-google-storage": "^1.3.0",
+    "partial": "0.0.7",
     "prettier": "^2.3.2",
     "prop-types": "^15.7.2",
     "tls": "0.0.1"
